move config utils to own file
